refactor(viewer): modify getCircleRadius location to Annotation
