+++ libs/insight-viewer/src/Annotation/utils/getCircleRadius.spec.ts
